25.Medical_Chat_bot/requirements.txt
langchain
langchain-core

langchain-openai
openai

python-dotenv

numpy
scikit-learn